requirements.txt
aiogram==3.17.0
aiohttp==3.11.11
beautifulsoup4==4.12.3
httpx==0.28.1
python-docx==1.1.2
python-dotenv==1.0.1
tzdata==2025.1
